[demo][hello] refactor directory structure
