+++ demo/hello/go.mod
@@ -1,3 +1,3 @@
-module hello
+module github.com/tbh26/g3/demo
 
 go 1.20
