+++ demo/hello/hello/hello.go
@@ -1,4 +1,4 @@
-package main
+package hello
 
 import (
 	"fmt"
